Refactor signin prompt into component
