+++ components/SaveForm/SaveForm.module.scss
@@ -38,14 +38,3 @@
   display: flex;
   flex-direction: row;
 }
-
-.upsell {
-  margin-top: 20vh;
-  text-align: center;
-
-  ul {
-    color: #666;
-    font-size: $font-size-base;
-    text-align: left;
-  }
-}
